refactor(canvas): complete chat ownership boundaries
