+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatConversationStatus/__tests__/ChatConversationStatus.test.tsx
@@ -2,8 +2,8 @@
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import { ChatConversationStatus } from '../ChatConversationStatus';
+import { I18nProvider } from '../../../../i18n';
+import { ChatConversationStatus } from '..';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
