+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatImageLightbox/__tests__/ChatImageLightbox.test.tsx
@@ -2,8 +2,8 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import { ChatImageLightbox, type LightboxImage } from '../ChatImageLightbox';
+import { I18nProvider } from '../../../../i18n';
+import { ChatImageLightbox, type LightboxImage } from '..';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
