+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatInput/__tests__/ChatInput.execution-attachments.test.tsx
@@ -2,8 +2,8 @@
 import { act, createRef } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import { ChatInput } from '../ChatInput';
+import { I18nProvider } from '../../../../i18n';
+import { ChatInput } from '..';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
