+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMentionPopup/__tests__/ChatMentionPopup.test.tsx
@@ -2,9 +2,9 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import { ChatMentionPopup } from '../ChatMentionPopup';
-import type { MentionItem } from '../../../types';
+import { I18nProvider } from '../../../../i18n';
+import { ChatMentionPopup } from '..';
+import type { MentionItem } from '../../../../types';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
