+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessage/ChatToolCalls/__tests__/ChatToolCallDetails.test.tsx
@@ -3,8 +3,8 @@ import { act, type ReactNode } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it } from 'vitest';
 
-import { I18nProvider } from '../../../i18n';
-import { ChatToolCallDetails } from '../ChatMessage/ChatToolCalls/ChatToolCallDetails';
+import { I18nProvider } from '../../../../../i18n';
+import { ChatToolCallDetails } from '../ChatToolCallDetails';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
