+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatMessage/index.css
@@ -518,23 +518,6 @@
   font-style: italic;
 }
 
-/* ─── Resize handle ──────────────────────────────────────── */
-
-.chat-panel-resize {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 4px;
-  height: 100%;
-  cursor: col-resize;
-  z-index: 10;
-}
-
-.chat-panel-resize:hover,
-.chat-panel-resize--active {
-  background: rgba(55, 53, 47, 0.12);
-}
-
 /* ─── Tool call indicators ───────────────────────────────── */
 
 .chat-message-body {
@@ -953,62 +936,6 @@
   text-overflow: ellipsis;
   white-space: nowrap;
   min-width: 0;
-}
-
-/* Back bar shown after a session jump (return to where the jump started). */
-.chat-session-back {
-  flex-shrink: 0;
-  padding: 5px 12px 0;
-}
-
-.chat-session-back-btn {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  width: 100%;
-  min-width: 0;
-  padding: 4px 8px;
-  border: none;
-  border-radius: var(--radius-sm);
-  background: transparent;
-  font-family: inherit;
-  font-size: 11.5px;
-  color: rgba(55, 53, 47, 0.5);
-  cursor: pointer;
-  transition: background 0.1s, color 0.1s;
-  text-align: left;
-}
-
-.chat-session-back-btn:hover:not(:disabled) {
-  background: rgba(120, 80, 200, 0.07);
-  color: rgba(120, 80, 200, 0.85);
-}
-
-.chat-session-back-btn:disabled {
-  cursor: progress;
-  opacity: 0.45;
-}
-
-.chat-session-back-icon {
-  flex-shrink: 0;
-  opacity: 0.7;
-}
-
-.chat-session-back-prefix {
-  flex-shrink: 0;
-  white-space: nowrap;
-}
-
-.chat-session-back-label {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  min-width: 0;
-  opacity: 0.65;
-}
-
-.chat-session-back-label::before {
-  content: '· ';
 }
 
 /* Inline session citation chip (assistant body, `@[session:...|label]`). */
