+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/__tests__/ChatPageResponsive.test.ts
@@ -1,9 +1,9 @@
 import { readFileSync } from 'node:fs';
 import { describe, expect, it } from 'vitest';
 
-const chatPageCss = readFileSync(new URL('../ChatPageBody/index.css', import.meta.url), 'utf8');
-const rightDockCss = readFileSync(new URL('../../dock/RightDock/index.css', import.meta.url), 'utf8');
-const rendererCss = readFileSync(new URL('../../../styles.css', import.meta.url), 'utf8');
+const chatPageCss = readFileSync(new URL('../index.css', import.meta.url), 'utf8');
+const rightDockCss = readFileSync(new URL('../../../dock/RightDock/index.css', import.meta.url), 'utf8');
+const rendererCss = readFileSync(new URL('../../../../styles.css', import.meta.url), 'utf8');
 
 describe('ChatPage narrow-window layout', () => {
   it('uses the chat container, not the full window, before the rail can squeeze the conversation', () => {
