+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatSessionsRail/__tests__/ChatSessionsRail.test.tsx
@@ -2,8 +2,8 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import { ChatSessionsRail, type UnifiedSession } from '../ChatSessionsRail';
+import { I18nProvider } from '../../../../i18n';
+import { ChatSessionsRail, type UnifiedSession } from '..';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
