+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatSessionsRail/index.css
@@ -391,3 +391,186 @@
   align-items: center;
   gap: 8px;
 }
+
+/* ─── Accessibility and responsive behavior ───────────────── */
+
+.chat-page-rail>.ui-textfield {
+  flex-shrink: 0;
+}
+
+.chat-page-rail-search {
+  height: 32px;
+  padding-inline: 9px;
+  border-color: var(--border);
+  border-radius: var(--radius-sm);
+  font-size: 12px;
+}
+
+.chat-page-rail-search::placeholder {
+  color: var(--chat-a11y-secondary);
+}
+
+.chat-page-rail-scroll {
+  max-height: none;
+}
+
+.chat-page-rail-new,
+.ui-btn.chat-page-rail-folder,
+.chat-page-rail-item {
+  min-height: 32px;
+  box-sizing: border-box;
+}
+
+.chat-page-rail-new:focus-visible,
+.chat-page-rail-item:focus-visible {
+  outline: 2px solid var(--accent);
+  outline-offset: -2px;
+}
+
+.chat-page-rail-folder-count,
+.chat-page-rail-empty,
+.chat-page-rail-loading {
+  color: var(--chat-a11y-secondary);
+}
+
+.chat-page-rail-item-row,
+.chat-page-rail-item-shell {
+  min-width: 0;
+  width: 100%;
+}
+
+.chat-page-rail-item-row {
+  animation: chat-page-rail-item-reveal 0.18s cubic-bezier(0.2, 0.75, 0.25, 1) both;
+}
+
+.chat-page-rail-item-row:nth-child(2) { animation-delay: 0.018s; }
+.chat-page-rail-item-row:nth-child(3) { animation-delay: 0.036s; }
+.chat-page-rail-item-row:nth-child(n + 4) { animation-delay: 0.054s; }
+
+@keyframes chat-page-rail-item-reveal {
+  from {
+    opacity: 0;
+    transform: translateY(-4px);
+  }
+
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.chat-page-rail-item-shell {
+  position: relative;
+  display: flex;
+  align-items: center;
+}
+
+.chat-page-rail-item-shell .chat-page-rail-item {
+  min-width: 0;
+  flex: 1;
+}
+
+.chat-page-rail-item-actions {
+  position: absolute;
+  top: 50%;
+  right: 3px;
+  display: flex;
+  align-items: center;
+  gap: 1px;
+  padding-left: 12px;
+  background: linear-gradient(90deg, transparent, var(--surface) 12px);
+  opacity: 0;
+  pointer-events: none;
+  transform: translateY(-50%);
+}
+
+.chat-page-rail-item-shell:hover .chat-page-rail-item-actions,
+.chat-page-rail-item-shell:focus-within .chat-page-rail-item-actions {
+  opacity: 1;
+  pointer-events: auto;
+}
+
+.chat-page-rail-item-action { color: var(--chat-a11y-secondary); }
+.chat-page-rail-item-action--danger { color: var(--error); }
+
+.chat-page-rail-item-editor,
+.chat-page-rail-item-confirm {
+  display: flex;
+  align-items: center;
+  gap: 3px;
+  width: 100%;
+  min-width: 0;
+  padding: 3px 3px 3px 20px;
+  box-sizing: border-box;
+}
+
+.chat-page-rail-item-editor>.ui-textfield {
+  flex: 1;
+  min-width: 0;
+}
+
+.chat-page-rail-item-editor .ui-textfield__control {
+  height: 28px;
+  padding-inline: 7px;
+  font-size: 12px;
+}
+
+.chat-page-rail-item-confirm {
+  flex-wrap: wrap;
+  padding-block: 5px;
+  color: var(--text);
+  font-size: 11px;
+}
+
+.chat-page-rail-item-confirm>span {
+  width: 100%;
+  color: var(--chat-a11y-secondary);
+}
+
+@container chat-page (max-width: 760px) {
+  .chat-page-rail { width: 100%; }
+}
+
+@container chat-page (max-width: 520px) {
+  .chat-page-rail {
+    width: min(220px, 42cqw);
+    padding-inline: 6px;
+  }
+
+  .chat-page-rail-item { padding-left: 16px; }
+}
+
+@media (hover: none) {
+  .chat-page-rail-folder-row--actionable .chat-page-rail-folder-count { opacity: 0; }
+
+  .chat-page-rail-folder-row--actionable .ui-btn.chat-page-rail-folder-new {
+    opacity: 1;
+    pointer-events: auto;
+  }
+
+  .chat-page-rail-item-actions {
+    position: static;
+    flex-shrink: 0;
+    padding-left: 0;
+    background: transparent;
+    opacity: 1;
+    pointer-events: auto;
+    transform: none;
+  }
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .chat-page-rail-folder-chevron,
+  .chat-page-rail-folder-count,
+  .ui-btn.chat-page-rail-folder-new,
+  .chat-page-rail-item,
+  .chat-page-rail-new,
+  .chat-page-rail-item-actions {
+    transition: none;
+  }
+
+  .chat-page-rail-item-running-spinner,
+  .chat-page-rail-item-row {
+    animation: none;
+  }
+}
