+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatView/__tests__/ChatView.focus-recovery.test.tsx
@@ -2,9 +2,9 @@
 import { act, createRef, type ComponentProps } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import type { AgentChatMessage } from '../../../types';
-import { ChatView } from '../ChatView';
+import { I18nProvider } from '../../../../i18n';
+import type { AgentChatMessage } from '../../../../types';
+import { ChatView } from '..';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
