+++ apps/canvas-workspace/src/renderer/src/components/chat/SessionBackBar/index.css
@@ -1,0 +1,54 @@
+.chat-session-back {
+  flex-shrink: 0;
+  padding: 5px 12px 0;
+}
+
+.chat-session-back-btn {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  width: 100%;
+  min-width: 0;
+  padding: 4px 8px;
+  border: none;
+  border-radius: var(--radius-sm);
+  background: transparent;
+  font-family: inherit;
+  font-size: 11.5px;
+  color: rgba(55, 53, 47, 0.5);
+  cursor: pointer;
+  transition: background 0.1s, color 0.1s;
+  text-align: left;
+}
+
+.chat-session-back-btn:hover:not(:disabled) {
+  background: rgba(120, 80, 200, 0.07);
+  color: rgba(120, 80, 200, 0.85);
+}
+
+.chat-session-back-btn:disabled {
+  cursor: progress;
+  opacity: 0.45;
+}
+
+.chat-session-back-icon {
+  flex-shrink: 0;
+  opacity: 0.7;
+}
+
+.chat-session-back-prefix {
+  flex-shrink: 0;
+  white-space: nowrap;
+}
+
+.chat-session-back-label {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  min-width: 0;
+  opacity: 0.65;
+}
+
+.chat-session-back-label::before {
+  content: '· ';
+}
