+++ apps/canvas-workspace/src/renderer/src/components/dock/RightDock/TabChatAction.tsx
@@ -4,7 +4,7 @@ import { useI18n } from '../../../i18n';
 import { SparklesIcon } from '../../icons';
 import { Button } from '../../ui';
 import type { ChatDeliveryReceipt } from '../../../agent-chat/target';
-import { useChatDeliveryNotifier } from '../../../agent-chat/target/useChatDeliveryNotifier';
+import { useChatDeliveryNotifier } from '../../shell/useChatDeliveryNotifier';
 
 interface Props {
   tab: AgentContextTabRef;
