+++ apps/canvas-workspace/src/renderer/src/components/shell/Workbench/usePreviewNodeActionBridge.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import type { CanvasNode } from '../../../types';
 import type { NodeReferenceEntry } from '../../dock/ReferenceDrawer/types';
 import type { ChatDeliveryReceipt } from '../../../agent-chat/target';
-import { useChatDeliveryNotifier } from '../../../agent-chat/target/useChatDeliveryNotifier';
+import { useChatDeliveryNotifier } from '../useChatDeliveryNotifier';
 import { getNodeDisplayLabel } from '../../../utils/nodeLabel';
 import {
   PREVIEW_EVICT_OPEN_EVENT,
